+++ tests/JWT/Action/VerifyIdToken/EmulatorTest.php
@@ -8,6 +8,7 @@ use Kreait\Firebase\JWT\Action\VerifyIdToken;
 use Kreait\Firebase\JWT\Action\VerifyIdToken\Handler;
 use Kreait\Firebase\JWT\Action\VerifyIdToken\WithLcobucciJWT;
 use Kreait\Firebase\JWT\InsecureToken;
+use Kreait\Firebase\JWT\Keys\StaticKeys;
 use Kreait\Firebase\JWT\Tests\Action\VerifyIdToken\TestCase;
 
 /**
@@ -28,6 +29,21 @@ final class EmulatorTest extends TestCase
     {
         $token = $this->createHandler()->handle(VerifyIdToken::withToken($this->token->withoutSignature()->idToken()));
 
+        $this->assertInstanceOf(InsecureToken::class, $token);
+    }
+
+    public function testItAcceptsATokenWithoutKidHeader(): void
+    {
+        $token = $this->createHandler()->handle(VerifyIdToken::withToken($this->token->withoutHeader('kid')->idToken()));
+
+        $this->assertInstanceOf(InsecureToken::class, $token);
+    }
+
+    public function testItAcceptsEmptyKeys(): void
+    {
+        $this->keys = StaticKeys::empty();
+
+        $token = $this->createHandler()->handle(VerifyIdToken::withToken($this->token->idToken()));
         $this->assertInstanceOf(InsecureToken::class, $token);
     }